pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pro-nlp-finalproject-nlp-08"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "pillow>=12.1.0",
    "pytest>=9.0.2",
    "torch==2.8.0",
    "transformers>=4.48.0,<5.0.0",
    "hydra-core>=1.3.2",
    "requests>=2.32.5",
    "uvicorn[standard]>=0.40.0",
    "pyinstaller>=6.18.0",
    "faiss-cpu>=1.13.2",
    "ipykernel>=7.1.0",
    "torchvision>=0.23.0",
    "einops>=0.8.2",
    "huggingface-hub[hf-xet]>=0.36.2",
    "openai>=2.16.0",
    "pymupdf>=1.26.7",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "black[jupyter]>=25.12.0",
    "ipykernel>=6.29.5",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.black]
line-length = 79           # 한 줄 최대 길이
target-version = ['py311'] # 프로젝트의 파이썬 버전
include = '\.(py|pyi|ipynb)$'       # 포맷팅할 파일 확장자
exclude = '''
/(
    \.git
    | \.venv
)/
'''

